Clear any cached config values from the .env file
We should purge the connection config values before checking whether the SQL credentials inserted by the user are valid or not.

The DatabaseManager class caches the values at run-time from the auto-generated .env when installing/creating a laravel project. This makes changing the config values on the fly redundant without purging the values first.

Tested on Laravel 5.8 and 6.0 and works brilliantly.