.env file saving proccess abstractization
Rather than returning a string (which can differ, based on the selected language, let's return a boolean instead.

We can then do the validation in the Controller based on the boolean value (true/false).

In its current state, if the .env file can't be written, the user still gets redirected to the database controller.

In the proposed form, we can potentially cancel the request, and return back with an error message, so the user doesn't lose his input values until the .env file permissions get updated.

The entire process will work exactly the same as before, just provides more flexibility for those with custom changes to it.